examples/sample/django/openapi/clients/python/cfg_support/pyproject.toml
[tool.poetry]
name = "django_revolution_cfg_support"
version = "1.0.0 (v1)"
description = "A client library for accessing Django CFG Sample API"
authors = []
readme = "README.md"
packages = [
    { include = "django_revolution_cfg_support" },
]
include = ["CHANGELOG.md", "django_revolution_cfg_support/py.typed"]

[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
